Módulo 16 - TdD IV - Pandas avançado/requirements.txt
streamlit-pandas-profiling
ydata-profiling
scikit-learn